+++ app/index.html
@@ -20,13 +20,13 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous">
     </script>
-  <script  src="/particles.js"></script>
+  <!-- <script  src="/particles.js"></script>
   <script >
     /* particlesJS.load(@dom-id, @path-json, @callback (optional)); */
     particlesJS.load('particle', '/particles.json', function () {
       console.log('callback - particles.js config loaded');
     });
-  </script>
+  </script> -->
 </body>
 
 </html>
